Factored out menu elements to separate components
